Rename and use inner CodeAction
